requirements.txt
Dash
flask
tensorflow
numpy
pybase64
opencv-python
Pillow
